+++ compiler/stack.c
@@ -925,7 +925,7 @@ int stack_calculate(int *value, struct stack_item *si, int q, int save_if_cannot
           b++;
         }
         else {
-          int priority = get_op_priority(si[k].value);
+          int priority = get_op_priority((int)si[k].value);
           
           b--;
           while (b != -1 && op[b] != SI_OP_LEFT && get_op_priority(op[b]) >= priority) {
